fix: вносит исправления по фидбэку
