+++ js/main.js
@@ -1,4 +1,3 @@
-import {createArrayPhotoItems} from './create-array-photo-items.js';
-import {PHOTO_QUANTITY} from './create-array-photo-items-config.js';
+import {createArrayPhotoItems, PHOTO_QUANTITY} from './data.js';
 
 window.console.log(createArrayPhotoItems(PHOTO_QUANTITY));
